requirements.txt
numpy==1.14.2
Pillow==5.1.0
tensorflow_gpu==1.7.0
